Exception Hendling

+++ API REST, camadas, CRUD, exceções, validações/crud-clients/src/main/java/com/moreira/crud_clients/services/ClientService.java
@@ -3,7 +3,8 @@ package com.moreira.crud_clients.services;
 import com.moreira.crud_clients.dto.ClientDTO;
 import com.moreira.crud_clients.entities.Client;
 import com.moreira.crud_clients.repositories.ClientRepository;
-import org.springframework.beans.BeanUtils;
+import com.moreira.crud_clients.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -26,7 +27,8 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id) {
-        Client result = clientRepository.findById(id).get();
+        Client result = clientRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ClientDTO(result);
     }
 
@@ -40,18 +42,25 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto) {
-        Client entity = clientRepository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = clientRepository.save(entity);
-        return new ClientDTO(entity);
+        try {
+            Client entity = clientRepository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = clientRepository.save(entity);
+            return new ClientDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
+    @Transactional
     public void delete(Long id) {
+        if (!clientRepository.existsById(id))
+            throw new ResourceNotFoundException("Recurso não encontrado");
+
         clientRepository.deleteById(id);
     }
 
     private void copyDtoToEntity(ClientDTO dto, Client entity) {
-        //BeanUtils.copyProperties(dto, entity);
         entity.setName(dto.getName());
         entity.setCpf(dto.getCpf());
         entity.setIncome(dto.getIncome());
